+++ CHANGES.md
@@ -16,6 +16,8 @@ Change Log
 * Fix `AsyncLoader` result bug
 * Remove `Terria.error` event handler
 * Refactor `workbench.add` to return `Result`
+* Consolidated network request / CORS error message - it is now in `t("core.terriaError.networkRequestMessage")`. It can be injected into other translation strings like so:
+  * `"groupNotAvailableMessage": "An error occurred while invoking the ArcGIS REST service.  \n\n$t(core.terriaError.networkRequestMessage)"`
 * [The next improvement]
 
 #### 8.1.2